feat: 패키지 경로 ecommerce -> member 변경
