--- user/src/main/java/com/example/ecommerce/MemberApplication.java
@@ -1,13 +1,0 @@
-package com.example.ecommerce;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MemberApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MemberApplication.class, args);
-	}
-
-}
