--- user/src/main/java/com/example/ecommerce/common/exception/ErrorResponse.java
@@ -1,6 +1,0 @@
-package com.example.ecommerce.common.exception;
-
-public record ErrorResponse(
-	String msg
-) {
-}
