--- user/src/main/java/com/example/ecommerce/common/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package com.example.ecommerce.common.exception;
-
-public class NotFoundException extends RuntimeException {
-
-	public NotFoundException(String message) {
-		super(message);
-	}
-}
