--- user/src/main/java/com/example/ecommerce/common/exception/UnauthorizedException.java
@@ -1,8 +1,0 @@
-package com.example.ecommerce.common.exception;
-
-public class UnauthorizedException extends RuntimeException {
-
-	public UnauthorizedException(String message) {
-		super(message);
-	}
-}
