--- user/src/main/java/com/example/ecommerce/member/application/dto/req/AddressDto.java
@@ -1,9 +1,0 @@
-package com.example.ecommerce.member.application.dto.req;
-
-public record AddressDto(
-	String state,
-	String street,
-	String city,
-	String zipCode
-) {
-}
