--- user/src/main/java/com/example/ecommerce/member/application/dto/req/PhoneNumberDto.java
@@ -1,8 +1,0 @@
-package com.example.ecommerce.member.application.dto.req;
-
-public record PhoneNumberDto(
-	String countryCode,
-	String areaCode,
-	String number
-) {
-}
