--- user/src/main/java/com/example/ecommerce/member/application/port/in/SignUpUseCase.java
@@ -1,8 +1,0 @@
-package com.example.ecommerce.member.application.port.in;
-
-import com.example.ecommerce.member.application.dto.req.SignUpReqDto;
-
-public interface SignUpUseCase {
-
-	Long signUp(SignUpReqDto reqDto);
-}
