+++ user/src/main/java/com/example/member/MemberApplication.java
@@ -1,0 +1,13 @@
+package com.example.member;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = {"com.example.auth", "com.example.member"})
+public class MemberApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MemberApplication.class, args);
+	}
+
+}
