+++ user/src/main/java/com/example/member/common/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package com.example.member.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+}
