+++ user/src/main/java/com/example/member/common/exception/ErrorResponse.java
@@ -1,0 +1,6 @@
+package com.example.member.common.exception;
+
+public record ErrorResponse(
+	String msg
+) {
+}
