+++ user/src/main/java/com/example/member/common/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package com.example.member.common.exception;
+
+public class NotFoundException extends RuntimeException {
+
+	public NotFoundException(String message) {
+		super(message);
+	}
+}
