+++ user/src/main/java/com/example/member/common/exception/UnauthorizedException.java
@@ -1,0 +1,8 @@
+package com.example.member.common.exception;
+
+public class UnauthorizedException extends RuntimeException {
+
+	public UnauthorizedException(String message) {
+		super(message);
+	}
+}
