+++ user/src/main/java/com/example/member/member/application/dto/req/AddressDto.java
@@ -1,0 +1,9 @@
+package com.example.member.member.application.dto.req;
+
+public record AddressDto(
+	String state,
+	String street,
+	String city,
+	String zipCode
+) {
+}
