+++ user/src/main/java/com/example/member/member/application/dto/req/LoginReqDto.java
@@ -1,0 +1,12 @@
+package com.example.member.member.application.dto.req;
+
+import jakarta.validation.constraints.NotNull;
+
+public record LoginReqDto(
+	@NotNull
+	String email,
+
+	@NotNull
+	String password
+) {
+}
