+++ user/src/main/java/com/example/member/member/application/dto/req/PhoneNumberDto.java
@@ -1,0 +1,8 @@
+package com.example.member.member.application.dto.req;
+
+public record PhoneNumberDto(
+	String countryCode,
+	String areaCode,
+	String number
+) {
+}
