+++ user/src/main/java/com/example/member/member/application/port/in/SignUpUseCase.java
@@ -1,0 +1,8 @@
+package com.example.member.member.application.port.in;
+
+import com.example.member.member.application.dto.req.SignUpReqDto;
+
+public interface SignUpUseCase {
+
+	Long signUp(SignUpReqDto reqDto);
+}
