+++ user/src/main/java/com/example/member/member/infrastructure/repository/MemberJpaRepository.java
@@ -1,0 +1,11 @@
+package com.example.member.member.infrastructure.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.member.member.infrastructure.entity.MemberEntity;
+
+public interface MemberJpaRepository extends JpaRepository<MemberEntity, Long> {
+	Optional<MemberEntity> findByEncryptEmail(String encryptEmail);
+}
